split streams vs io resolves #9
